[SharedCache] Process the .symbols cache file
A significant number of symbols are not being defined because there is currently no support for parsing the .symbols cache file. This commit adds that support.

`m_symbolInfos` has been modified to be a vector of references to `Symbol`s which is similar to [this PR](https://github.com/Vector35/binaryninja-api/pull/6197). It makes more sense for the use case where `m_symbolInfos` is used as a symbol cache, otherwise a bunch of time is spent transforming between the old style of `m_symbolInfos` entries to Binary Ninja `Symbol`s.

This commit does require a metadata version bump. I felt this was necessary to determine which symbols to load from the symbols cache. The problem is that the `m_images` container does not store the images in the order they are found in the DSC. The index they are at determines the location of their symbols in the symbols cache file. Rather than converting `m_images` to a vector and relying on its ordering being correct, it seemed more prudent to store the index of the image in the `CacheImage` structure. As this is serialized, the metadata version has to be bumped to accomodate the change.