[SharedCache] UI improvements for loading images and processing Objective-C
This commit adds a number of new buttons to the DSC triage view images tab. The user can now select and load multiple images as well as their direct dependencies. Additionally they can choose whether Objective-C processing is done on image load, or manually trigger it for a given image later on.

In a future commit it would be good to add the ability to change the depth of the dependencies to load for the selected images and tell the user how many images it will cause to be loaded.